For now, use a git dependency on cassandra-sys-rs.
Required so we can depend on our own forked version of the
low-level binding.

Will be restored to a versioned dependency once our
cassandra-sys-rs is on crates.io.

Because cassandra-sys-rs no longer exports a symbol (which we do not
in fact use), we should no longer export it either.
See https://github.com/Metaswitch/cassandra-sys-rs/pull/2 and
https://datastax-oss.atlassian.net/browse/CPP-473 for details.

Signed-off-by: Keith Wansbrough <Keith.Wansbrough@metaswitch.com>